Extract transition_map into its own files.
